Datetime, date and time stands for universal times.
Local-time library is not well suited for this, so we can remove it
now. By using universal time we delegate to other librarys to provide
good datetime functions to the user. As well as we delegate to the
lisp implementation to handle with timezones.

We are not going to cover it local and UTC times by the moment.

It should be easy to modify format-value and parse-value to accept
them in a future.